refactor: decoupling DevPocketTools component
