+++ src/views/dev-pocket-tools/composables/useClipboardStatus.ts
@@ -1,0 +1,36 @@
+import { ref } from 'vue'
+
+import type { CopyStatus, OutputTone } from '../types'
+
+export function useClipboardStatus() {
+  const copyStatus = ref<CopyStatus | null>(null)
+  let copyTimer = 0
+
+  function setCopyStatus(key: string, state: OutputTone, message: string) {
+    copyStatus.value = { key, state, message }
+    window.clearTimeout(copyTimer)
+    copyTimer = window.setTimeout(() => {
+      copyStatus.value = null
+    }, 2200)
+  }
+
+  async function copyText(key: string, value: string) {
+    if (!value) {
+      setCopyStatus(key, 'error', 'Chưa có dữ liệu để copy.')
+      return
+    }
+
+    try {
+      await navigator.clipboard.writeText(value)
+      setCopyStatus(key, 'success', 'Đã copy vào clipboard.')
+    } catch {
+      setCopyStatus(key, 'error', 'Trình duyệt không cho phép copy tự động.')
+    }
+  }
+
+  return {
+    copyStatus,
+    copyText,
+    setCopyStatus,
+  }
+}
