+++ src/views/dev-pocket-tools/index.vue
@@ -1,1055 +1,70 @@
 <script setup lang="ts">
-import { computed, ref } from 'vue'
 import { RouterLink } from 'vue-router'
 
-type QueryParam = {
-  key: string
-  value: string
-}
-
-type RegexMatch = {
-  index: number
-  text: string
-  groups: string[]
-}
-
-type JwtTimeField = {
-  key: string
-  raw: string
-  formatted: string
-}
-
-type CaseVariant = {
-  label: string
-  value: string
-}
-
-type DiffLine = {
-  type: 'same' | 'added' | 'removed' | 'changed'
-  left: string
-  right: string
-}
-
-type ParsedHeader = {
-  key: string
-  value: string
-}
-
-type CsvRow = Record<string, string>
-type ToolPageId = 'builders' | 'data' | 'time' | 'inspect'
-
-type JsonPrimitive = string | number | boolean | null
-type JsonValue = JsonPrimitive | JsonValue[] | { [key: string]: JsonValue }
-
-type OutputTone = 'default' | 'success' | 'error'
-
-type CopyStatus = {
-  key: string
-  state: OutputTone
-  message: string
-}
-
-const jsonInput = ref('{\n  "name": "J2TEAM",\n  "tools": ["json", "base64", "regex"],\n  "active": true\n}')
-const base64Text = ref('Xin chào J2TEAM')
-const base64Value = ref('WGluIGNow6BvIEoyVEVBTQ==')
-const urlInput = ref('https://example.com/search?q=vibe%20code&lang=vi&sort=newest')
-const timestampInput = ref(Math.floor(Date.now() / 1000).toString())
-const dateTimeInput = ref(formatDateTimeLocal(new Date()))
-const regexPattern = ref('\\b(dev|tool)\\b')
-const regexFlags = ref('gi')
-const regexSample = ref('Dev Pocket Tools giúp dev kiểm tra regex ngay trong trình duyệt.')
-const slugInput = ref('Bộ công cụ bỏ túi dành cho lập trình viên Việt Nam')
-const jwtInput = ref(
-  'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpXVCBEZWNvZGVyIiwiaWF0IjoxNzQxMzA0MjAwLCJleHAiOjE3NDEzMDc4MDB9.signature',
-)
-const hashInput = ref('Dev Pocket Tools')
-const hashAlgorithm = ref('SHA-256')
-const hashOutput = ref('')
-const hashStatus = ref<{ tone: OutputTone; message: string }>({
-  tone: 'default',
-  message: 'Nhập text và sinh hash bằng Web Crypto API.',
-})
-const caseInput = ref('dev pocket tools v2')
-const jsonTypeInput = ref('{\n  "id": 1,\n  "name": "J2TEAM",\n  "tags": ["tools", "dev"],\n  "profile": {\n    "active": true,\n    "score": 9.8\n  }\n}')
-const uuidCount = ref(3)
-const uuidList = ref<string[]>([])
-const htmlInput = ref('<button title="Dev & Tools">Click me</button>')
-const diffLeft = ref('alpha\nbeta\ngamma\ndelta')
-const diffRight = ref('alpha\nbeta updated\ngamma\nepsilon')
-const cronInput = ref('*/15 9-18 * * 1-5')
-const queryBuilderInput = ref('q=dev tools\nlang=vi\nsort=newest')
-const numberBaseInput = ref('255')
-const numberBase = ref<'10' | '2' | '8' | '16'>('10')
-const relativeTimestampInput = ref(Math.floor(Date.now() / 1000).toString())
-const rawHeadersInput = ref('Content-Type: application/json\nAuthorization: Bearer token\nX-Request-ID: abc-123')
-const csvInput = ref('name,role,active\nJ2TEAM,community,true\nKhiem,author,true')
-const colorInput = ref('#ff6b4a')
-const copyStatus = ref<CopyStatus | null>(null)
-const currentPage = ref<ToolPageId>('builders')
-const toolPages = [
-  { id: 'builders', label: 'Builders', description: 'JWT, hash, case, type, UUID' },
-  { id: 'data', label: 'Data & Text', description: 'JSON, Base64, URL, query, HTML' },
-  { id: 'time', label: 'Time & Rules', description: 'Timestamp, relative, regex, cron, slug' },
-  { id: 'inspect', label: 'Inspect & Convert', description: 'Diff, headers, CSV, color, number base' },
-] satisfies Array<{ id: ToolPageId; label: string; description: string }>
-const defaultToolPage = toolPages[0]!
-const currentToolPage = computed(() => toolPages.find((page) => page.id === currentPage.value) ?? defaultToolPage)
-
-const jsonResult = computed(() => {
-  const source = jsonInput.value.trim()
-
-  if (!source) {
-    return { tone: 'default' as const, message: 'Dán JSON thô vào đây để format hoặc kiểm tra cú pháp.', output: '' }
-  }
-
-  try {
-    const parsed: object = JSON.parse(source) as object
-    return { tone: 'success' as const, message: 'JSON hợp lệ.', output: JSON.stringify(parsed, null, 2) }
-  } catch (error) {
-    return {
-      tone: 'error' as const,
-      message: error instanceof Error ? error.message : 'Không thể phân tích JSON.',
-      output: '',
-    }
-  }
-})
-
-const urlResult = computed(() => {
-  const source = urlInput.value.trim()
-
-  if (!source) {
-    return {
-      tone: 'default' as const,
-      encoded: '',
-      decoded: '',
-      origin: '',
-      pathname: '',
-      queryParams: [] as QueryParam[],
-      message: 'Nhập URL đầy đủ hoặc chỉ phần query để encode, decode và parse.',
-    }
-  }
-
-  const encoded = encodeURIComponent(source)
-  let decoded = source
-
-  try {
-    decoded = decodeURIComponent(source)
-  } catch {
-    decoded = 'Chuỗi hiện tại không phải dữ liệu URL-encoded hợp lệ để decode toàn phần.'
-  }
-
-  try {
-    const url = normalizeUrlInput(source)
-    return {
-      tone: 'success' as const,
-      encoded,
-      decoded,
-      origin: url.origin,
-      pathname: `${url.pathname}${url.hash}`,
-      queryParams: Array.from(url.searchParams.entries()).map(([key, value]) => ({ key, value })),
-      message: url.searchParams.size ? `Đã parse ${url.searchParams.size} query param.` : 'URL hợp lệ nhưng không có query param.',
-    }
-  } catch (error) {
-    return {
-      tone: 'error' as const,
-      encoded,
-      decoded,
-      origin: '',
-      pathname: '',
-      queryParams: [] as QueryParam[],
-      message: error instanceof Error ? error.message : 'Không thể phân tích URL.',
-    }
-  }
-})
-
-const timestampResult = computed(() => {
-  const source = timestampInput.value.trim()
-
-  if (!source) {
-    return { tone: 'default' as const, normalized: '', iso: '', local: '', message: 'Nhập Unix timestamp theo giây hoặc mili giây.' }
-  }
-
-  const numericValue = Number(source)
-  if (!Number.isFinite(numericValue)) {
-    return { tone: 'error' as const, normalized: '', iso: '', local: '', message: 'Timestamp phải là số hợp lệ.' }
-  }
-
-  const milliseconds = source.length <= 10 ? numericValue * 1000 : numericValue
-  const date = new Date(milliseconds)
-  if (Number.isNaN(date.getTime())) {
-    return { tone: 'error' as const, normalized: '', iso: '', local: '', message: 'Timestamp nằm ngoài phạm vi ngày giờ hợp lệ.' }
-  }
-
-  return {
-    tone: 'success' as const,
-    normalized: Math.floor(milliseconds / 1000).toString(),
-    iso: date.toISOString(),
-    local: date.toLocaleString('vi-VN'),
-    message: source.length <= 10 ? 'Đang đọc dưới dạng giây.' : 'Đang đọc dưới dạng mili giây.',
-  }
-})
-
-const dateTimeToTimestamp = computed(() => {
-  if (!dateTimeInput.value) {
-    return { tone: 'default' as const, seconds: '', milliseconds: '', message: 'Chọn ngày giờ để đổi ngược sang timestamp.' }
-  }
-
-  const date = new Date(dateTimeInput.value)
-  if (Number.isNaN(date.getTime())) {
-    return { tone: 'error' as const, seconds: '', milliseconds: '', message: 'Ngày giờ không hợp lệ.' }
-  }
-
-  return {
-    tone: 'success' as const,
-    seconds: Math.floor(date.getTime() / 1000).toString(),
-    milliseconds: date.getTime().toString(),
-    message: 'Đã chuyển từ ngày giờ cục bộ sang Unix timestamp.',
-  }
-})
-
-const regexResult = computed(() => {
-  if (!regexPattern.value) {
-    return { tone: 'default' as const, message: 'Nhập pattern và flags để kiểm tra regex.', matches: [] as RegexMatch[] }
-  }
-
-  try {
-    const expression = new RegExp(regexPattern.value, regexFlags.value)
-    const matches = collectRegexMatches(expression, regexSample.value)
-    return {
-      tone: matches.length ? ('success' as const) : ('default' as const),
-      message: matches.length ? `Tìm thấy ${matches.length} match.` : 'Không có đoạn nào khớp.',
-      matches,
-    }
-  } catch (error) {
-    return {
-      tone: 'error' as const,
-      message: error instanceof Error ? error.message : 'Regex không hợp lệ.',
-      matches: [] as RegexMatch[],
-    }
-  }
-})
-
-const slugResult = computed(() => {
-  const source = slugInput.value.trim()
-  if (!source) {
-    return { tone: 'default' as const, slug: '', message: 'Nhập tiêu đề hoặc cụm từ để tạo slug.' }
-  }
-
-  const slug = toSlug(source)
-  return {
-    tone: slug ? ('success' as const) : ('error' as const),
-    slug,
-    message: slug ? 'Slug đã sẵn sàng để dùng trong URL.' : 'Không thể tạo slug từ nội dung hiện tại.',
-  }
-})
-
-const jwtResult = computed(() => {
-  const source = jwtInput.value.trim()
-
-  if (!source) {
-    return {
-      tone: 'default' as const,
-      message: 'Dán JWT để decode header và payload.',
-      header: '',
-      payload: '',
-      timeFields: [] as JwtTimeField[],
-    }
-  }
-
-  const segments = source.split('.')
-
-  if (segments.length < 2) {
-    return {
-      tone: 'error' as const,
-      message: 'JWT phải có ít nhất 2 phần: header.payload.',
-      header: '',
-      payload: '',
-      timeFields: [] as JwtTimeField[],
-    }
-  }
-
-  try {
-    const headerText = decodeBase64UrlToText(segments[0] ?? '')
-    const payloadText = decodeBase64UrlToText(segments[1] ?? '')
-    const headerJson = JSON.parse(headerText) as JsonValue
-    const payloadJson = JSON.parse(payloadText) as JsonValue
-
-    return {
-      tone: 'success' as const,
-      message: 'Đã decode JWT. Signature chưa được verify trong v2.',
-      header: JSON.stringify(headerJson, null, 2),
-      payload: JSON.stringify(payloadJson, null, 2),
-      timeFields: extractJwtTimeFields(payloadJson),
-    }
-  } catch (error) {
-    return {
-      tone: 'error' as const,
-      message: error instanceof Error ? error.message : 'Không thể decode JWT.',
-      header: '',
-      payload: '',
-      timeFields: [] as JwtTimeField[],
-    }
-  }
-})
-
-const caseVariants = computed(() => {
-  const words = splitWords(caseInput.value)
-
-  return [
-    { label: 'camelCase', value: toCamelCase(words) },
-    { label: 'snake_case', value: words.join('_') },
-    { label: 'kebab-case', value: words.join('-') },
-    { label: 'PascalCase', value: toPascalCase(words) },
-    { label: 'CONSTANT_CASE', value: words.join('_').toUpperCase() },
-  ] satisfies CaseVariant[]
-})
-
-const jsonTypeResult = computed(() => {
-  const source = jsonTypeInput.value.trim()
-
-  if (!source) {
-    return {
-      tone: 'default' as const,
-      message: 'Nhập JSON mẫu để sinh interface TypeScript.',
-      output: '',
-    }
-  }
-
-  try {
-    const parsed = JSON.parse(source) as JsonValue
-    return {
-      tone: 'success' as const,
-      message: 'Đã sinh TypeScript từ JSON mẫu.',
-      output: generateTypeScriptType('RootPayload', parsed),
-    }
-  } catch (error) {
-    return {
-      tone: 'error' as const,
-      message: error instanceof Error ? error.message : 'JSON không hợp lệ.',
-      output: '',
-    }
-  }
-})
-
-const htmlEscapeResult = computed(() => {
-  const source = htmlInput.value
-
-  if (!source) {
-    return {
-      tone: 'default' as const,
-      escaped: '',
-      unescaped: '',
-      message: 'Nhập HTML hoặc text để escape và unescape.',
-    }
-  }
-
-  const escaped = escapeHtml(source)
-
-  return {
-    tone: 'success' as const,
-    escaped,
-    unescaped: unescapeHtml(source),
-    message: 'Đã chuyển đổi giữa dạng raw và escaped.',
-  }
-})
-
-const diffResult = computed(() => {
-  const leftLines = diffLeft.value.split('\n')
-  const rightLines = diffRight.value.split('\n')
-  const total = Math.max(leftLines.length, rightLines.length)
-  const lines: DiffLine[] = []
-
-  for (let index = 0; index < total; index += 1) {
-    const left = leftLines[index] ?? ''
-    const right = rightLines[index] ?? ''
-
-    if (left === right) {
-      lines.push({ type: 'same', left, right })
-    } else if (!left && right) {
-      lines.push({ type: 'added', left: '', right })
-    } else if (left && !right) {
-      lines.push({ type: 'removed', left, right: '' })
-    } else {
-      lines.push({ type: 'changed', left, right })
-    }
-  }
-
-  return {
-    lines,
-    summary: `${lines.filter((line) => line.type !== 'same').length} dòng có khác biệt.`,
-  }
-})
-
-const cronResult = computed(() => {
-  const source = cronInput.value.trim()
-
-  if (!source) {
-    return {
-      tone: 'default' as const,
-      message: 'Nhập cron 5 field theo định dạng phút giờ ngày-tháng tháng thứ.',
-      output: '',
-    }
-  }
-
-  const parts = source.split(/\s+/)
-
-  if (parts.length !== 5) {
-    return {
-      tone: 'error' as const,
-      message: 'Cron v3 hiện chỉ hỗ trợ đúng 5 field.',
-      output: '',
-    }
-  }
-
-  const [minute = '*', hour = '*', dayOfMonth = '*', month = '*', dayOfWeek = '*'] = parts
-
-  return {
-    tone: 'success' as const,
-    message: 'Diễn giải cron ở mức đọc nhanh, không thay thế parser đầy đủ.',
-    output: `Chạy ${describeCronField(minute, 'phút')} | ${describeCronField(hour, 'giờ')} | ${describeCronField(dayOfMonth, 'ngày trong tháng')} | ${describeCronField(month, 'tháng')} | ${describeCronField(dayOfWeek, 'thứ trong tuần')}.`,
-  }
-})
-
-const queryBuilderResult = computed(() => {
-  const entries = queryBuilderInput.value
-    .split('\n')
-    .map((line) => line.trim())
-    .filter(Boolean)
-    .map((line) => {
-      const separatorIndex = line.indexOf('=')
-
-      if (separatorIndex === -1) {
-        return { key: line, value: '' }
-      }
-
-      return {
-        key: line.slice(0, separatorIndex).trim(),
-        value: line.slice(separatorIndex + 1).trim(),
-      }
-    })
-    .filter((entry) => entry.key)
-
-  const params = new URLSearchParams()
-
-  entries.forEach((entry) => {
-    params.append(entry.key, entry.value)
-  })
-
-  return {
-    entries,
-    query: params.toString(),
-    preview: params.toString() ? `?${params.toString()}` : '',
-  }
-})
-
-const numberBaseResult = computed(() => {
-  const source = numberBaseInput.value.trim()
-
-  if (!source) {
-    return {
-      tone: 'default' as const,
-      message: 'Nhập số và chọn hệ cơ số nguồn.',
-      decimal: '',
-      binary: '',
-      octal: '',
-      hex: '',
-    }
-  }
-
-  const parsed = Number.parseInt(source, Number(numberBase.value))
-
-  if (Number.isNaN(parsed)) {
-    return {
-      tone: 'error' as const,
-      message: 'Giá trị không hợp lệ cho hệ cơ số đã chọn.',
-      decimal: '',
-      binary: '',
-      octal: '',
-      hex: '',
-    }
-  }
-
-  return {
-    tone: 'success' as const,
-    message: 'Đã chuyển đổi giữa các hệ cơ số phổ biến.',
-    decimal: parsed.toString(10),
-    binary: parsed.toString(2),
-    octal: parsed.toString(8),
-    hex: parsed.toString(16).toUpperCase(),
-  }
-})
-
-const relativeTimeResult = computed(() => {
-  const source = relativeTimestampInput.value.trim()
-
-  if (!source) {
-    return {
-      tone: 'default' as const,
-      message: 'Nhập Unix timestamp để xem relative time.',
-      absolute: '',
-      relative: '',
-    }
-  }
-
-  const numericValue = Number(source)
-
-  if (!Number.isFinite(numericValue)) {
-    return {
-      tone: 'error' as const,
-      message: 'Timestamp phải là số hợp lệ.',
-      absolute: '',
-      relative: '',
-    }
-  }
-
-  const milliseconds = source.length <= 10 ? numericValue * 1000 : numericValue
-  const date = new Date(milliseconds)
-
-  if (Number.isNaN(date.getTime())) {
-    return {
-      tone: 'error' as const,
-      message: 'Timestamp không hợp lệ.',
-      absolute: '',
-      relative: '',
-    }
-  }
-
-  return {
-    tone: 'success' as const,
-    message: 'Đã tính relative time theo thời điểm hiện tại của máy bạn.',
-    absolute: date.toLocaleString('vi-VN'),
-    relative: formatRelativeTime(date.getTime() - Date.now()),
-  }
-})
-
-const parsedHeaders = computed(() => {
-  const lines = rawHeadersInput.value
-    .split('\n')
-    .map((line) => line.trim())
-    .filter(Boolean)
-
-  const headers = lines
-    .map((line) => {
-      const separatorIndex = line.indexOf(':')
-
-      if (separatorIndex === -1) {
-        return null
-      }
-
-      return {
-        key: line.slice(0, separatorIndex).trim(),
-        value: line.slice(separatorIndex + 1).trim(),
-      }
-    })
-    .filter((header): header is ParsedHeader => header !== null)
-
-  return {
-    headers,
-    json: headers.length ? JSON.stringify(Object.fromEntries(headers.map((header) => [header.key, header.value])), null, 2) : '',
-  }
-})
-
-const csvResult = computed(() => {
-  const source = csvInput.value.trim()
-
-  if (!source) {
-    return {
-      tone: 'default' as const,
-      message: 'Dán CSV có header ở dòng đầu để convert sang JSON.',
-      output: '',
-      rowCount: 0,
-    }
-  }
-
-  try {
-    const rows = parseCsv(source)
-
-    if (rows.length < 2) {
-      return {
-        tone: 'error' as const,
-        message: 'CSV cần ít nhất một dòng header và một dòng dữ liệu.',
-        output: '',
-        rowCount: 0,
-      }
-    }
-
-    const [headers = [], ...dataRows] = rows
-    const normalizedHeaders = headers.map((header, index) => header || `column_${index + 1}`)
-    const objects = dataRows.map((row) => csvRowToObject(normalizedHeaders, row))
-
-    return {
-      tone: 'success' as const,
-      message: `Đã convert ${objects.length} dòng CSV sang JSON.`,
-      output: JSON.stringify(objects, null, 2),
-      rowCount: objects.length,
-    }
-  } catch (error) {
-    return {
-      tone: 'error' as const,
-      message: error instanceof Error ? error.message : 'Không thể parse CSV.',
-      output: '',
-      rowCount: 0,
-    }
-  }
-})
-
-const colorResult = computed(() => {
-  const normalized = normalizeHexColor(colorInput.value.trim())
-
-  if (!normalized) {
-    return {
-      tone: 'error' as const,
-      message: 'Nhập màu dạng HEX như #FF6B4A hoặc #F64.',
-      hex: '',
-      rgb: '',
-      hsl: '',
-    }
-  }
-
-  const { r, g, b } = hexToRgb(normalized)
-  const { h, s, l } = rgbToHsl(r, g, b)
-
-  return {
-    tone: 'success' as const,
-    message: 'Đã convert màu sang RGB và HSL.',
-    hex: normalized.toUpperCase(),
-    rgb: `rgb(${r}, ${g}, ${b})`,
-    hsl: `hsl(${h}, ${s}%, ${l}%)`,
-  }
-})
-
-function formatJson() {
-  if (jsonResult.value.output) jsonInput.value = jsonResult.value.output
-}
-
-function encodeBase64() {
-  base64Value.value = encodeTextToBase64(base64Text.value)
-}
-
-function decodeBase64() {
-  try {
-    base64Text.value = decodeBase64ToText(base64Value.value)
-  } catch (error) {
-    setCopyStatus('base64', 'error', error instanceof Error ? error.message : 'Base64 không hợp lệ.')
-  }
-}
-
-async function copyText(key: string, value: string) {
-  if (!value) {
-    setCopyStatus(key, 'error', 'Chưa có dữ liệu để copy.')
-    return
-  }
-
-  try {
-    await navigator.clipboard.writeText(value)
-    setCopyStatus(key, 'success', 'Đã copy vào clipboard.')
-  } catch {
-    setCopyStatus(key, 'error', 'Trình duyệt không cho phép copy tự động.')
-  }
-}
-
-function setTimestampNow() {
-  timestampInput.value = Math.floor(Date.now() / 1000).toString()
-  dateTimeInput.value = formatDateTimeLocal(new Date())
-}
-
-function useConvertedTimestamp() {
-  if (dateTimeToTimestamp.value.seconds) timestampInput.value = dateTimeToTimestamp.value.seconds
-}
-
-function useDecodedUrl() {
-  if (urlResult.value.decoded && !urlResult.value.decoded.startsWith('Chuỗi hiện tại')) urlInput.value = urlResult.value.decoded
-}
-
-async function generateHash() {
-  const source = hashInput.value
-
-  if (!source) {
-    hashOutput.value = ''
-    hashStatus.value = {
-      tone: 'error',
-      message: 'Nhập nội dung trước khi sinh hash.',
-    }
-    return
-  }
-
-  if (!('crypto' in window) || !('subtle' in window.crypto)) {
-    hashOutput.value = ''
-    hashStatus.value = {
-      tone: 'error',
-      message: 'Trình duyệt hiện tại không hỗ trợ Web Crypto API.',
-    }
-    return
-  }
-
-  try {
-    const digest = await window.crypto.subtle.digest(hashAlgorithm.value, new TextEncoder().encode(source))
-    hashOutput.value = Array.from(new Uint8Array(digest))
-      .map((byte) => byte.toString(16).padStart(2, '0'))
-      .join('')
-    hashStatus.value = {
-      tone: 'success',
-      message: `Đã sinh ${hashAlgorithm.value}.`,
-    }
-  } catch (error) {
-    hashOutput.value = ''
-    hashStatus.value = {
-      tone: 'error',
-      message: error instanceof Error ? error.message : 'Không thể sinh hash.',
-    }
-  }
-}
-
-function generateUuids() {
-  const total = Math.min(Math.max(uuidCount.value, 1), 20)
-  uuidCount.value = total
-  uuidList.value = Array.from({ length: total }, () => window.crypto.randomUUID())
-}
-
-let copyTimer = 0
-
-function setCopyStatus(key: string, state: OutputTone, message: string) {
-  copyStatus.value = { key, state, message }
-  window.clearTimeout(copyTimer)
-  copyTimer = window.setTimeout(() => {
-    copyStatus.value = null
-  }, 2200)
-}
-
-function normalizeUrlInput(source: string) {
-  if (source.startsWith('?')) return new URL(source, 'https://example.dev')
-  try {
-    return new URL(source)
-  } catch {
-    return new URL(source.startsWith('/') ? source : `/${source}`, 'https://example.dev')
-  }
-}
-
-function collectRegexMatches(expression: RegExp, sample: string) {
-  if (!sample) return [] as RegexMatch[]
-  if (expression.global) {
-    return Array.from(sample.matchAll(expression)).map((match) => ({
-      index: match.index ?? 0,
-      text: match[0],
-      groups: match.slice(1).filter((group): group is string => typeof group === 'string'),
-    }))
-  }
-
-  const firstMatch = expression.exec(sample)
-  if (!firstMatch) return [] as RegexMatch[]
-
-  return [{ index: firstMatch.index ?? 0, text: firstMatch[0], groups: firstMatch.slice(1).filter((group): group is string => typeof group === 'string') }]
-}
-
-function encodeTextToBase64(value: string) {
-  const bytes = new TextEncoder().encode(value)
-  let binary = ''
-  bytes.forEach((byte) => {
-    binary += String.fromCharCode(byte)
-  })
-  return btoa(binary)
-}
-
-function decodeBase64ToText(value: string) {
-  const binary = atob(value.trim())
-  const bytes = Uint8Array.from(binary, (character) => character.charCodeAt(0))
-  return new TextDecoder().decode(bytes)
-}
-
-function decodeBase64UrlToText(value: string) {
-  const normalized = value.replace(/-/g, '+').replace(/_/g, '/')
-  const padding = normalized.length % 4 === 0 ? '' : '='.repeat(4 - (normalized.length % 4))
-
-  return decodeBase64ToText(`${normalized}${padding}`)
-}
-
-function escapeHtml(value: string) {
-  return value
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;')
-    .replace(/'/g, '&#39;')
-}
-
-function unescapeHtml(value: string) {
-  return value
-    .replace(/&lt;/g, '<')
-    .replace(/&gt;/g, '>')
-    .replace(/&quot;/g, '"')
-    .replace(/&#39;/g, "'")
-    .replace(/&amp;/g, '&')
-}
-
-function extractJwtTimeFields(value: JsonValue) {
-  if (!isJsonObject(value)) {
-    return [] as JwtTimeField[]
-  }
-
-  return ['iat', 'exp', 'nbf']
-    .map((key) => {
-      const rawValue = value[key]
-
-      if (typeof rawValue !== 'number') {
-        return null
-      }
-
-      return {
-        key,
-        raw: rawValue.toString(),
-        formatted: new Date(rawValue * 1000).toLocaleString('vi-VN'),
-      }
-    })
-    .filter((field): field is JwtTimeField => field !== null)
-}
-
-function toSlug(value: string) {
-  return value
-    .normalize('NFD')
-    .replace(/[\u0300-\u036f]/g, '')
-    .replace(/đ/g, 'd')
-    .replace(/Đ/g, 'd')
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, '-')
-    .replace(/^-+|-+$/g, '')
-}
-
-function formatDateTimeLocal(date: Date) {
-  const year = date.getFullYear()
-  const month = String(date.getMonth() + 1).padStart(2, '0')
-  const day = String(date.getDate()).padStart(2, '0')
-  const hours = String(date.getHours()).padStart(2, '0')
-  const minutes = String(date.getMinutes()).padStart(2, '0')
-  return `${year}-${month}-${day}T${hours}:${minutes}`
-}
-
-function splitWords(value: string) {
-  return value
-    .replace(/([a-z0-9])([A-Z])/g, '$1 $2')
-    .normalize('NFD')
-    .replace(/[\u0300-\u036f]/g, '')
-    .replace(/đ/g, 'd')
-    .replace(/Đ/g, 'D')
-    .replace(/[^A-Za-z0-9]+/g, ' ')
-    .trim()
-    .split(/\s+/)
-    .filter(Boolean)
-    .map((word) => word.toLowerCase())
-}
-
-function toCamelCase(words: string[]) {
-  return words
-    .map((word, index) => (index === 0 ? word : `${word.charAt(0).toUpperCase()}${word.slice(1)}`))
-    .join('')
-}
-
-function toPascalCase(words: string[]) {
-  return words.map((word) => `${word.charAt(0).toUpperCase()}${word.slice(1)}`).join('')
-}
-
-function isJsonObject(value: JsonValue): value is { [key: string]: JsonValue } {
-  return typeof value === 'object' && value !== null && !Array.isArray(value)
-}
-
-function jsonValueToType(value: JsonValue, indentLevel: number): string {
-  if (value === null) {
-    return 'null'
-  }
-
-  if (Array.isArray(value)) {
-    if (!value.length) {
-      return 'never[]'
-    }
-
-    const itemTypes = Array.from(new Set(value.map((item) => jsonValueToType(item, indentLevel)))).join(' | ')
-    return `(${itemTypes})[]`
-  }
-
-  if (isJsonObject(value)) {
-    const indent = '  '.repeat(indentLevel)
-    const childIndent = '  '.repeat(indentLevel + 1)
-    const lines = Object.entries(value).map(([key, child]) => `${childIndent}${formatPropertyKey(key)}: ${jsonValueToType(child, indentLevel + 1)}`)
-
-    return `{\n${lines.join(';\n')};\n${indent}}`
-  }
-
-  if (typeof value === 'string') {
-    return 'string'
-  }
-
-  if (typeof value === 'number') {
-    return 'number'
-  }
-
-  return 'boolean'
-}
-
-function formatPropertyKey(key: string) {
-  return /^[A-Za-z_$][A-Za-z0-9_$]*$/.test(key) ? key : `'${key}'`
-}
-
-function generateTypeScriptType(name: string, value: JsonValue) {
-  const typeName = toPascalCase(splitWords(name)) || 'GeneratedType'
-  const body = jsonValueToType(value, 0)
-
-  if (isJsonObject(value)) {
-    return `interface ${typeName} ${body}`
-  }
-
-  return `type ${typeName} = ${body}`
-}
-
-function describeCronField(value: string, label: string) {
-  if (value === '*') {
-    return `mỗi ${label}`
-  }
-
-  if (value.startsWith('*/')) {
-    return `mỗi ${value.slice(2)} ${label}`
-  }
-
-  if (value.includes(',')) {
-    return `${label} thuộc các giá trị ${value}`
-  }
-
-  if (value.includes('-')) {
-    return `${label} trong khoảng ${value}`
-  }
-
-  return `${label} = ${value}`
-}
-
-function formatRelativeTime(diffMs: number) {
-  const formatter = new Intl.RelativeTimeFormat('vi', { numeric: 'auto' })
-  const minutes = Math.round(diffMs / 60000)
-  const hours = Math.round(diffMs / 3600000)
-  const days = Math.round(diffMs / 86400000)
-
-  if (Math.abs(days) >= 1) {
-    return formatter.format(days, 'day')
-  }
-
-  if (Math.abs(hours) >= 1) {
-    return formatter.format(hours, 'hour')
-  }
-
-  return formatter.format(minutes, 'minute')
-}
-
-function parseCsv(source: string) {
-  const rows: string[][] = []
-  let currentCell = ''
-  let currentRow: string[] = []
-  let inQuotes = false
-
-  for (let index = 0; index < source.length; index += 1) {
-    const character = source[index] ?? ''
-    const nextCharacter = source[index + 1] ?? ''
-
-    if (character === '"') {
-      if (inQuotes && nextCharacter === '"') {
-        currentCell += '"'
-        index += 1
-      } else {
-        inQuotes = !inQuotes
-      }
-      continue
-    }
-
-    if (character === ',' && !inQuotes) {
-      currentRow.push(currentCell)
-      currentCell = ''
-      continue
-    }
-
-    if ((character === '\n' || character === '\r') && !inQuotes) {
-      if (character === '\r' && nextCharacter === '\n') {
-        index += 1
-      }
-
-      currentRow.push(currentCell)
-      rows.push(currentRow)
-      currentCell = ''
-      currentRow = []
-      continue
-    }
-
-    currentCell += character
-  }
-
-  if (inQuotes) {
-    throw new Error('CSV có dấu nháy kép chưa đóng.')
-  }
-
-  currentRow.push(currentCell)
-  rows.push(currentRow)
-
-  return rows.filter((row) => row.some((cell) => cell.length > 0))
-}
-
-function csvRowToObject(headers: string[], row: string[]) {
-  return headers.reduce<CsvRow>((result, header, index) => {
-    result[header] = row[index] ?? ''
-    return result
-  }, {})
-}
-
-function normalizeHexColor(value: string) {
-  if (/^#?[0-9a-fA-F]{3}$/.test(value)) {
-    const raw = value.replace('#', '')
-    return `#${raw.split('').map((char) => `${char}${char}`).join('')}`
-  }
-
-  if (/^#?[0-9a-fA-F]{6}$/.test(value)) {
-    return value.startsWith('#') ? value : `#${value}`
-  }
-
-  return ''
-}
-
-function hexToRgb(hex: string) {
-  const raw = hex.replace('#', '')
-
-  return {
-    r: Number.parseInt(raw.slice(0, 2), 16),
-    g: Number.parseInt(raw.slice(2, 4), 16),
-    b: Number.parseInt(raw.slice(4, 6), 16),
-  }
-}
-
-function rgbToHsl(r: number, g: number, b: number) {
-  const red = r / 255
-  const green = g / 255
-  const blue = b / 255
-  const max = Math.max(red, green, blue)
-  const min = Math.min(red, green, blue)
-  const lightness = (max + min) / 2
-  const delta = max - min
-
-  if (delta === 0) {
-    return { h: 0, s: 0, l: Math.round(lightness * 100) }
-  }
-
-  const saturation = delta / (1 - Math.abs(2 * lightness - 1))
-  let hue = 0
-
-  if (max === red) {
-    hue = ((green - blue) / delta) % 6
-  } else if (max === green) {
-    hue = (blue - red) / delta + 2
-  } else {
-    hue = (red - green) / delta + 4
-  }
-
-  return {
-    h: Math.round(hue * 60 < 0 ? hue * 60 + 360 : hue * 60),
-    s: Math.round(saturation * 100),
-    l: Math.round(lightness * 100),
-  }
-}
-
-generateHash()
-generateUuids()
+import { useDevPocketTools } from './composables/useDevPocketTools'
+
+const {
+  base64Text,
+  base64Value,
+  caseInput,
+  caseVariants,
+  colorInput,
+  colorResult,
+  copyText,
+  cronInput,
+  cronResult,
+  csvInput,
+  csvResult,
+  currentPage,
+  currentToolPage,
+  dateTimeInput,
+  dateTimeToTimestamp,
+  decodeBase64,
+  diffLeft,
+  diffResult,
+  diffRight,
+  encodeBase64,
+  formatJson,
+  generateHash,
+  generateUuids,
+  hashAlgorithm,
+  hashInput,
+  hashOutput,
+  hashStatus,
+  htmlEscapeResult,
+  htmlInput,
+  jsonInput,
+  jsonResult,
+  jsonTypeInput,
+  jsonTypeResult,
+  jwtInput,
+  jwtResult,
+  numberBase,
+  numberBaseInput,
+  numberBaseResult,
+  parsedHeaders,
+  queryBuilderInput,
+  queryBuilderResult,
+  rawHeadersInput,
+  regexFlags,
+  regexPattern,
+  regexResult,
+  regexSample,
+  relativeTimeResult,
+  relativeTimestampInput,
+  setTimestampNow,
+  slugInput,
+  slugResult,
+  timestampInput,
+  timestampResult,
+  toolPages,
+  urlInput,
+  urlResult,
+  useConvertedTimestamp,
+  useDecodedUrl,
+  uuidCount,
+  uuidList,
+} = useDevPocketTools()
 </script>
 
 <template>
